test(contracts): 32kb transfer, 9MM gas

+++ packages/contracts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L1ERC20Gateway.sol
@@ -89,7 +89,7 @@ contract OVM_L1ERC20Gateway is Abs_L1TokenGateway {
      */
     function _handleFinalizeWithdrawal(
         address _to,
-        uint _amount
+        uint256 _amount
     )
         internal
         override
